drivers: hse: HSE Hash driver support
Introduce support for digest algorithms MD5, SHA1, and
SHA{2,3}-{224, 256, 384, 512} and in the HSE hash driver. To avoid
performance loss, instead of sending one block at a time (sending all of
them at the same is impossible, the inputs are too big for OP-TEE to
handle), a safety net was introduced, sending a fixed size of data a
time, and the remaining data in the final operation. This way, for
smaller blocks the throughtput is ~2 times better, and for bigger blocks
~16 times better. Not choosing this buffered approach would result in a
driver slower than the previous software approach.

Issue: ALB-8816, ALB-10654
Signed-off-by: David Anghel <david.anghel@nxp.com>
Signed-off-by: Bogdan Roman <bogdan-gabriel.roman@nxp.com>