drivers: hse: add support for ECDSA in the HSE driver
Introduce support for ECDSA in the HSE driver. This driver is set by
default to the value "no", since it removes the support for ECDH by
removing the software implementation.

To use the elliptic curves vectors, which are under an MIT license, we
are required to add the creator's copyright in the header.

Issue: ALB-9957, ALB-10654
Signed-off-by: David Anghel <david.anghel@nxp.com>
Signed-off-by: Bogdan Roman <bogdan-gabriel.roman@nxp.com>