drivers: hse: add support for managing keys in the HSE driver
Keys in HSE are kept in slots and the slots are grouped into keygroups,
each keygroup having a particular type (i.e. AES, HMAC, etc), number
of slots and ID. Keygroups are grouped into catalogs and there are two
catalogs which can be formatted by the user: RAM and NVM Catalogs.

In the HSE Driver, the keygroups are  kept in a list data structure
Each entry in the list represents a keygroup and there's a unique key
group per key type in the list.

To make use of a key, one must first acquire a slot and then import the
key material into that slot. There are separate functions for each
action, but, for convenience, there is also a third function,
hse_acquire_and_import_key() which does both actions.
The key slot acquisition returns a key handle. The key handle is the
unique identifier of a key slot in HSE's internal RAM/NVM Catalogs and
is used in service requests which need to reference the key for a
specific operation (i.e. encryption with a symmetric key).

When the operation is done and the key is no longer needed, one must
release and erase the key. Again, hse_release_and_erase_key() has
been implemented to satisfy these actions.

Issue: ALB-8890, ALB-9302, ALB-10654
Signed-off-by: Bogdan Roman <bogdan-gabriel.roman@nxp.com>